pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-idit"
version = "0.1.0"
description = "Your life, hash-linked and signed. A personal blockchain for sovereign documentation."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Lot's Wife Project", email = "hello@lotswife.org"},
]
keywords = ["blockchain", "personal", "documentation", "signing", "immutable", "autobiography"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Archiving",
]
dependencies = [
    "pynacl>=1.5.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[project.scripts]
idit = "idit.cli:main"

[project.urls]
Homepage = "https://lotswife.org"
Repository = "https://github.com/idit-life/personal-idit"
"Future Network" = "https://idit.life"

[tool.setuptools.packages.find]
include = ["idit*"]
